Rename common to peripherals and move
